move beacon block and blob to eth2/types
